// src/api/schema.ts
export type ID = string
export type GitObjectID = string
export type DateTime = string
export type JSONCString = string

export interface IGraphQLResponseRoot {
    data?: IQuery | IMutation
    errors?: IGraphQLResponseError[]
}

export interface IGraphQLResponseError {
    /** Required for all errors */
    message: string
    locations?: IGraphQLResponseErrorLocation[]
    /** 7.2.2 says 'GraphQL servers may provide additional entries to error' */
    [propName: string]: any
}

export interface IGraphQLResponseErrorLocation {
    line: number
    column: number
}

export interface IQuery {
    __typename: 'Query'

    /**
     * @deprecated "this will be removed."
     */
    root: IQuery
    node: Node | null
    campaigns: ICampaignConnection
    repository: IRepository | null
    repositoryRedirect: RepositoryRedirect | null
    externalServices: IExternalServiceConnection
    repositories: IRepositoryConnection
    phabricatorRepo: IPhabricatorRepo | null
    currentUser: IUser | null
    user: IUser | null
    users: IUserConnection
    organization: IOrg | null
    organizations: IOrgConnection
    renderMarkdown: string
    highlightCode: string
    settingsSubject: SettingsSubject | null
    viewerSettings: ISettingsCascade

    /**
     * @deprecated "use viewerSettings instead"
     */
    viewerConfiguration: IConfigurationCascade
    clientConfiguration: IClientConfigurationDetails
    searchFilterSuggestions: ISearchFilterSuggestions
    search: ISearch | null
    savedSearches: ISavedSearch[]
    repoGroups: IRepoGroup[]
    site: ISite
    surveyResponses: ISurveyResponseConnection
    extensionRegistry: IExtensionRegistry
    dotcom: IDotcomQuery
    statusMessages: StatusMessage[]
    namespace: Namespace | null
    authorizedUserRepositories: IRepositoryConnection
    usersWithPendingPermissions: string[]
}
// ...
export type Node =
    | ICampaign
    | IPatchSet
    | IPatch
    | IRepository
    | IGitCommit
    | IUser
    | IOrg
    | IOrganizationInvitation
    | IAccessToken
    | IExternalAccount
    | IExternalService
    | IGitRef
    | ILSIFUpload
    | IExternalChangeset
    | IChangesetEvent
    | ISavedSearch
    | IRegistryExtension
    | IProductSubscription
    | IProductLicense

export interface INode {
    __typename: 'Node'
    id: ID
}

export enum CampaignState {
    OPEN = 'OPEN',
    CLOSED = 'CLOSED',
}

export interface ICampaignConnection {
    __typename: 'CampaignConnection'
    nodes: ICampaign[]
    totalCount: number
    pageInfo: IPageInfo
}
// ...
export interface IRepository {
    __typename: 'Repository'
    id: ID
    name: string

    /**
     * @deprecated "Use name."
     */
    uri: string
    description: string
    language: string
    createdAt: DateTime
    updatedAt: DateTime | null
    commit: IGitCommit | null
    mirrorInfo: IMirrorRepositoryInfo
    externalRepository: IExternalRepository | null
    isFork: boolean
    isArchived: boolean
    isPrivate: boolean
    externalServices: IExternalServiceConnection

    /**
     * @deprecated "use Repository.mirrorInfo.cloneInProgress instead"
     */
    cloneInProgress: boolean
    textSearchIndex: IRepositoryTextSearchIndex | null
    url: string
    externalURLs: IExternalLink[]
    defaultBranch: IGitRef | null
    gitRefs: IGitRefConnection
    branches: IGitRefConnection
    tags: IGitRefConnection
    comparison: IRepositoryComparison
    contributors: IRepositoryContributorConnection

    /**
     * @deprecated "use repositoryRedirect query instead"
     */
    redirectURL: string | null
    viewerCanAdminister: boolean
    icon: string
    label: IMarkdown
    detail: IMarkdown
    matches: ISearchResultMatch[]
    lsifUploads: ILSIFUploadConnection
    authorizedUsers: IUserConnection
    permissionsInfo: IPermissionsInfo | null
}
// ...
export interface IUser {
    __typename: 'User'
    id: ID
    username: string

    /**
     * @deprecated "use emails instead"
     */
    email: string
    displayName: string | null
    avatarURL: string | null
    url: string
    settingsURL: string | null
    createdAt: DateTime
    updatedAt: DateTime | null
    siteAdmin: boolean
    builtinAuth: boolean
    latestSettings: ISettings | null
    settingsCascade: ISettingsCascade

    /**
     * @deprecated "Use settingsCascade instead. This field is a deprecated alias for it and will be removed in a future release."
     */
    configurationCascade: IConfigurationCascade
    organizations: IOrgConnection
    organizationMemberships: IOrganizationMembershipConnection
    tags: string[]
    usageStatistics: IUserUsageStatistics
    eventLogs: IEventLogsConnection
    emails: IUserEmail[]
    accessTokens: IAccessTokenConnection
    externalAccounts: IExternalAccountConnection
    session: ISession
    viewerCanAdminister: boolean
    viewerCanChangeUsername: boolean
    surveyResponses: ISurveyResponse[]
    urlForSiteAdminBilling: string | null
    databaseID: number
    namespaceName: string
}

export interface IEventLogsOnUserArguments {
    first?: number | null
}

export interface IAccessTokensOnUserArguments {
    first?: number | null
}

export interface IExternalAccountsOnUserArguments {
    first?: number | null
}

export type SettingsSubject = IUser | IOrg | ISite | IDefaultSettings
// ...
export interface ISettingsCascade {
    __typename: 'SettingsCascade'
    subjects: SettingsSubject[]
    final: string

    /**
     * @deprecated "use final instead"
     */
    merged: IConfiguration
}

export interface IConfigurationCascade {
    __typename: 'ConfigurationCascade'

    /**
     * @deprecated "use SettingsCascade.subjects instead"
     */
    subjects: SettingsSubject[]

    /**
     * @deprecated "use SettingsCascade.final instead"
     */
    merged: IConfiguration
}

export type Namespace = IUser | IOrg

export interface INamespace {
    __typename: 'Namespace'
    id: ID
    namespaceName: string
    url: string
}

export interface IOrgConnection {
    __typename: 'OrgConnection'
    nodes: IOrg[]
    totalCount: number
}

export interface IOrg {
    __typename: 'Org'
    id: ID
    name: string
    displayName: string | null
    createdAt: DateTime
    members: IUserConnection
    latestSettings: ISettings | null
    settingsCascade: ISettingsCascade

    /**
     * @deprecated "Use settingsCascade instead. This field is a deprecated alias for it and will be removed in a future release."
     */
    configurationCascade: IConfigurationCascade
    viewerPendingInvitation: IOrganizationInvitation | null
    viewerCanAdminister: boolean
    viewerIsMember: boolean
    url: string
    settingsURL: string | null
    namespaceName: string
}

export interface IUserConnection {
    __typename: 'UserConnection'
    nodes: IUser[]
    totalCount: number
    pageInfo: IPageInfo
}
// ...
export interface ISymbol {
    __typename: 'Symbol'
    name: string
    containerName: string | null
    kind: SymbolKind
    language: string
    location: ILocation
    url: string
    canonicalURL: string
    fileLocal: boolean
}
// ...
export interface IGitBlob {
    __typename: 'GitBlob'
    path: string
    name: string
    isDirectory: boolean
    content: string
    binary: boolean
    richHTML: string
    commit: IGitCommit
    repository: IRepository
    url: string
    canonicalURL: string
    externalURLs: IExternalLink[]
    blame: IHunk[]
    highlight: IHighlightedFile
    submodule: ISubmodule | null
    symbols: ISymbolConnection
    isSingleChild: boolean
    lsif: ILSIFQueryResolver | null
}
// ...
export interface IExternalServiceConnection {
    __typename: 'ExternalServiceConnection'
    nodes: IExternalService[]
    totalCount: number
    pageInfo: IPageInfo
}
// ...
export type GitRevSpec = IGitRef | IGitRevSpecExpr | IGitObject
// ...
export type RepositoryRedirect = IRepository | IRedirect

export interface IRedirect {
    __typename: 'Redirect'
    url: string
}

export enum RepositoryOrderBy {
    REPOSITORY_NAME = 'REPOSITORY_NAME',
    REPO_CREATED_AT = 'REPO_CREATED_AT',
    REPOSITORY_CREATED_AT = 'REPOSITORY_CREATED_AT',
}

export interface IRepositoryConnection {
    __typename: 'RepositoryConnection'
    nodes: IRepository[]
    totalCount: number | null
    pageInfo: IPageInfo
}

export interface ITotalCountOnRepositoryConnectionArguments {
    /**
     * @default false
     */
    precise?: boolean | null
}

export interface IPhabricatorRepo {
    __typename: 'PhabricatorRepo'
    name: string

    /**
     * @deprecated "use name instead"
     */
    uri: string
    callsign: string
    url: string
}

export enum UserActivePeriod {
    TODAY = 'TODAY',
    THIS_WEEK = 'THIS_WEEK',
    THIS_MONTH = 'THIS_MONTH',
    ALL_TIME = 'ALL_TIME',
}

export interface IMarkdownOptions {
    alwaysNil?: string | null
}

export interface IClientConfigurationDetails {
    __typename: 'ClientConfigurationDetails'
    contentScriptUrls: string[]
    parentSourcegraph: IParentSourcegraphDetails
}

export interface IParentSourcegraphDetails {
    __typename: 'ParentSourcegraphDetails'
    url: string
}

export interface ISearchFilterSuggestions {
    __typename: 'SearchFilterSuggestions'
    repogroup: string[]
    repo: string[]
}

export enum SearchVersion {
    V1 = 'V1',
    V2 = 'V2',
}

export enum SearchPatternType {
    literal = 'literal',
    regexp = 'regexp',
    structural = 'structural',
}

export interface ISearch {
    __typename: 'Search'
    results: ISearchResults
    suggestions: SearchSuggestion[]
    stats: ISearchResultsStats
}
// ...
export interface IFileMatch {
    __typename: 'FileMatch'
    file: IGitBlob
    repository: IRepository
    revSpec: GitRevSpec | null

    /**
     * @deprecated "use the file field instead"
     */
    resource: string
    symbols: ISymbol[]
    lineMatches: ILineMatch[]
    limitHit: boolean
}

export interface ILineMatch {
    __typename: 'LineMatch'
    preview: string
    lineNumber: number
    offsetAndLengths: number[][]
    limitHit: boolean
}
// ...
export interface ISavedSearch {
    __typename: 'SavedSearch'
    id: ID
    description: string
    query: string
    notify: boolean
    notifySlack: boolean
    namespace: Namespace
    slackWebhookURL: string | null
}

export interface IRepoGroup {
    __typename: 'RepoGroup'
    name: string
    repositories: string[]
}
// ...
export interface ISurveyResponseConnection {
    __typename: 'SurveyResponseConnection'
    nodes: ISurveyResponse[]
    totalCount: number
    last30DaysCount: number
    averageScore: number
    netPromoterScore: number
}

export interface IExtensionRegistry {
    __typename: 'ExtensionRegistry'
    extension: IRegistryExtension | null
    extensions: IRegistryExtensionConnection
    publishers: IRegistryPublisherConnection
    viewerPublishers: RegistryPublisher[]
    localExtensionIDPrefix: string | null
}
// ...
export interface IMutation {
    __typename: 'Mutation'
    createChangesets: IExternalChangeset[]
    addChangesetsToCampaign: ICampaign
    createCampaign: ICampaign
    createPatchSetFromPatches: IPatchSet
    updateCampaign: ICampaign
    retryCampaign: ICampaign
    deleteCampaign: IEmptyResponse | null
    closeCampaign: ICampaign
    publishCampaign: ICampaign
    publishChangeset: IEmptyResponse
    syncChangeset: IEmptyResponse
    updateUser: IEmptyResponse
    createOrganization: IOrg
    updateOrganization: IOrg
    deleteOrganization: IEmptyResponse | null
    addExternalService: IExternalService
    updateExternalService: IExternalService
    deleteExternalService: IEmptyResponse

    /**
     * @deprecated "update external service exclude setting."
     */
    setRepositoryEnabled: IEmptyResponse | null
    checkMirrorRepositoryConnection: ICheckMirrorRepositoryConnectionResult
    updateMirrorRepository: IEmptyResponse

    /**
     * @deprecated "syncer ensures all repositories are up to date."
     */
    updateAllMirrorRepositories: IEmptyResponse
    createUser: ICreateUserResult
    randomizeUserPassword: IRandomizeUserPasswordResult
    addUserEmail: IEmptyResponse
    removeUserEmail: IEmptyResponse
    setUserEmailVerified: IEmptyResponse
    deleteUser: IEmptyResponse | null
    updatePassword: IEmptyResponse | null
    createAccessToken: ICreateAccessTokenResult
    deleteAccessToken: IEmptyResponse
    deleteExternalAccount: IEmptyResponse
    inviteUserToOrganization: IInviteUserToOrganizationResult
    respondToOrganizationInvitation: IEmptyResponse
    resendOrganizationInvitationNotification: IEmptyResponse
    revokeOrganizationInvitation: IEmptyResponse
    addUserToOrganization: IEmptyResponse
    removeUserFromOrganization: IEmptyResponse | null
    setTag: IEmptyResponse
    addPhabricatorRepo: IEmptyResponse | null
    resolvePhabricatorDiff: IGitCommit | null

    /**
     * @deprecated "use logEvent instead"
     */
    logUserEvent: IEmptyResponse | null
    logEvent: IEmptyResponse | null
    sendSavedSearchTestNotification: IEmptyResponse | null
    settingsMutation: ISettingsMutation | null

    /**
     * @deprecated "use settingsMutation instead"
     */
    configurationMutation: ISettingsMutation | null
    updateSiteConfiguration: boolean
    setUserIsSiteAdmin: IEmptyResponse | null
    reloadSite: IEmptyResponse | null
    submitSurvey: IEmptyResponse | null
    requestTrial: IEmptyResponse | null
    extensionRegistry: IExtensionRegistryMutation
    dotcom: IDotcomMutation
    createSavedSearch: ISavedSearch
    updateSavedSearch: ISavedSearch
    deleteSavedSearch: IEmptyResponse | null
    deleteLSIFUpload: IEmptyResponse | null
    setRepositoryPermissionsForUsers: IEmptyResponse
}